Move across deposit fetching into separate package
